third_party/WebKit/LayoutTests/fast/events/wheel/wheelevent-document-createevent.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that document.createEvent() properly constructs and initializes a WheelEvent.");
var wheelEvent = document.createEvent("WheelEvent");

shouldBe("wheelEvent.deltaX", "0");
shouldBe("wheelEvent.deltaY", "0");
shouldBe("wheelEvent.deltaZ", "0");
shouldBe("wheelEvent.deltaMode", "0");

</script>
</body>
</html>
