third_party/blink/web_tests/platform/generic/fast/events/wheel/wheelevent-document-createevent-expected.txt

Tests that document.createEvent() properly constructs and initializes a WheelEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS wheelEvent.deltaX is 0
PASS wheelEvent.deltaY is 0
PASS wheelEvent.deltaZ is 0
PASS wheelEvent.deltaMode is 0
PASS successfullyParsed is true

TEST COMPLETE
